# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jambites/vendors/{vendorId}/menu:
    get:
      operationId: getVendorMenu
      tags: [jambites]
      summary: Get vendor menu
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vendor menu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorMenu"
  /jambites/orders:
    post:
      operationId: placeOrder
      tags: [jambites]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderBody"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /jambites/orders/{orderId}:
    get:
      operationId: getOrder
      tags: [jambites]
      summary: Get order status
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    Vendor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        rating:
          type: number
        deliveryTimeMin:
          type: integer
        deliveryFee:
          type: number
        distance:
          type: number
        isOpen:
          type: boolean
        category:
          type: string
        imageUrl:
          type: string
      required:
        - id
        - name
        - description
        - rating
        - deliveryTimeMin
        - deliveryFee
        - distance
        - isOpen
        - category
    MenuItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        isVeg:
          type: boolean
        isBestseller:
          type: boolean
        imageUrl:
          type: string
      required:
        - id
        - name
        - description
        - price
        - category
        - isVeg
        - isBestseller
    MenuSection:
      type: object
      properties:
        title:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/MenuItem"
      required:
        - title
        - items
    VendorMenu:
      type: object
      properties:
        vendor:
          $ref: "#/components/schemas/Vendor"
        sections:
          type: array
          items:
            $ref: "#/components/schemas/MenuSection"
      required:
        - vendor
        - sections
    CartItem:
      type: object
      properties:
        menuItemId:
          type: string
        quantity:
          type: integer
      required:
        - menuItemId
        - quantity
    PlaceOrderBody:
      type: object
      properties:
        vendorId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        deliveryLat:
          type: number
        deliveryLng:
          type: number
        landmark:
          type: string
        carNumber:
          type: string
        paymentMethod:
          type: string
      required:
        - vendorId
        - items
        - deliveryLat
        - deliveryLng
        - carNumber
        - paymentMethod
    Order:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        vendorId:
          type: string
        vendorName:
          type: string
        estimatedMinutes:
          type: integer
        totalAmount:
          type: number
        riderName:
          type: string
        riderPhone:
          type: string
        riderLat:
          type: number
        riderLng:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - vendorId
        - vendorName
        - estimatedMinutes
        - totalAmount
        - createdAt
...