# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: string;
  status: string;
  vendorId: string;
  vendorName: string;
  estimatedMinutes: number;
  totalAmount: number;
  riderName?: string;
  riderPhone?: string;
  riderLat?: number;
  riderLng?: number;
  createdAt: Date;
}