# lib/api-zod/src/generated/types/placeOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from "./cartItem";

export interface PlaceOrderBody {
  vendorId: string;
  items: CartItem[];
  deliveryLat: number;
  deliveryLng: number;
  landmark?: string;
  carNumber: string;
  paymentMethod: string;
}