# lib/api-zod/src/generated/types/vendor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Vendor {
  id: string;
  name: string;
  description: string;
  rating: number;
  deliveryTimeMin: number;
  deliveryFee: number;
  distance: number;
  isOpen: boolean;
  category: string;
  imageUrl?: string;
}