# lib/api-zod/src/generated/types/vendorMenu.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MenuSection } from "./menuSection";
import type { Vendor } from "./vendor";

export interface VendorMenu {
  vendor: Vendor;
  sections: MenuSection[];
}